+++ edison-jobs/src/main/java/de/otto/edison/jobs/service/JobRunnable.java
@@ -1,7 +1,6 @@
 package de.otto.edison.jobs.service;
 
 import de.otto.edison.jobs.definition.JobDefinition;
-import de.otto.edison.jobs.domain.JobInfo;
 import de.otto.edison.jobs.eventbus.JobEventPublisher;
 
 /**
@@ -22,8 +21,7 @@ public interface JobRunnable {
     /**
      * Executes the background task of the job and updates the JobInfo during execution.
      *
-     * @param jobInfo        the information about the job in execution.
      * @param jobEventPublisher publishes events to the event bus.
      */
-    void execute(JobInfo jobInfo, JobEventPublisher jobEventPublisher);
+    void execute(JobEventPublisher jobEventPublisher);
 }